Evaluation/Comparison_Other_Tools/LLM_powered_ML_code_smell_static_detection_3_versions/requirements.txt
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.1
